Refactor code structure for backends Fix date issue with logging
